requirements.txt
requests==2.31.0
bs4==0.0.2
python-dotenv==1.0.1
secure-smtplib
yagmail
